+++ assets/javascripts/discourse/connectors/topic-above-post-stream/encrypt-topic-class.gjs
@@ -1,0 +1,7 @@
+import bodyClass from "discourse/helpers/body-class";
+
+<template>
+  {{#if @outletArgs.model.encrypted_title}}
+    {{bodyClass "encrypted-topic-page"}}
+  {{/if}}
+</template>
